# docs/api/openapi.yaml
components:
  schemas:
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          title: Detail
          type: array
      title: HTTPValidationError
      type: object
...
    TopicCreate:
      description: 'Schema for creating a new topic.


        Used for API requests when creating a new topic.

        Slug is optional and will be auto-generated from title if not provided.'
      properties:
        aliases:
          items:
            type: string
          title: Aliases
          type: array
        description:
          anyOf:
          - type: string
          - type: 'null'
          title: Description
        external_refs:
          additionalProperties:
            type: string
          title: External Refs
          type: object
        identifier:
          anyOf:
          - maxLength: 255
            minLength: 1
            type: string
          - type: 'null'
          description: Topic CURIE
          title: Identifier
        slug:
          anyOf:
          - maxLength: 255
            minLength: 1
            type: string
          - type: 'null'
          title: Slug
        status:
          default: draft
          enum:
          - draft
          - active
          - deprecated
          - merged
          title: Status
          type: string
        taxonomy_id:
          format: uuid
          title: Taxonomy Id
          type: string
        taxonomy_identifier:
          anyOf:
          - maxLength: 255
            minLength: 1
            type: string
          - type: 'null'
          description: Parent taxonomy CURIE
          title: Taxonomy Identifier
        title:
          maxLength: 255
          minLength: 1
          title: Title
          type: string
      required:
      - taxonomy_id
      - title
      title: TopicCreate
      type: object
...
    TopicListResponse:
      description: Schema for paginated topic list responses.
      properties:
        items:
          description: List of topics
          items:
            $ref: '#/components/schemas/TopicResponse'
          title: Items
          type: array
        limit:
          description: Items per page
          minimum: 1.0
          title: Limit
          type: integer
        offset:
          description: Offset from start
          minimum: 0.0
          title: Offset
          type: integer
        total:
          description: Total count of topics
          minimum: 0.0
          title: Total
          type: integer
      required:
      - items
      - total
      - limit
      - offset
      title: TopicListResponse
      type: object
...
    TopicResponse:
      description: 'Schema for topic responses.


        Used for API responses. Includes all fields including ID and timestamps.'
      properties:
        aliases:
          description: Alternative names
          items:
            type: string
          title: Aliases
          type: array
        created_at:
          description: Creation timestamp
          format: date-time
          title: Created At
          type: string
        description:
          anyOf:
          - type: string
          - type: 'null'
          description: Optional topic description
          title: Description
        external_refs:
          additionalProperties:
            type: string
          description: External system references
          title: External Refs
          type: object
        id:
          description: Unique identifier
          format: uuid
          title: Id
          type: string
        identifier:
          description: Topic CURIE
          maxLength: 255
          minLength: 1
          title: Identifier
          type: string
        path:
          anyOf:
          - type: string
          - type: 'null'
          description: Materialized path for hierarchy
          title: Path
        slug:
          description: URL-friendly slug
          maxLength: 255
          minLength: 1
          title: Slug
          type: string
        status:
          default: draft
          description: Topic status
          enum:
          - draft
          - active
          - deprecated
          - merged
          title: Status
          type: string
        taxonomy_id:
          description: Parent taxonomy ID
          format: uuid
          title: Taxonomy Id
          type: string
        taxonomy_identifier:
          description: Parent taxonomy CURIE
          maxLength: 255
          minLength: 1
          title: Taxonomy Identifier
          type: string
        title:
          description: Topic title
          maxLength: 255
          minLength: 1
          title: Title
          type: string
        updated_at:
          description: Last update timestamp
          format: date-time
          title: Updated At
          type: string
      required:
      - taxonomy_id
      - taxonomy_identifier
      - identifier
      - title
      - slug
      - id
      - created_at
      - updated_at
      title: TopicResponse
      type: object
    TopicSearchRequest:
      description: Schema for topic search requests.
      properties:
        limit:
          default: 50
          description: Max results
          maximum: 1000.0
          minimum: 1.0
          title: Limit
          type: integer
        offset:
          default: 0
          description: Result offset
          minimum: 0.0
          title: Offset
          type: integer
        query:
          description: Search query
          minLength: 1
          title: Query
          type: string
        status:
          anyOf:
          - enum:
            - draft
            - active
            - deprecated
            - merged
            type: string
          - type: 'null'
          description: Filter by status
          title: Status
        taxonomy_id:
          anyOf:
          - format: uuid
            type: string
          - type: 'null'
          description: Filter by taxonomy
          title: Taxonomy Id
      required:
      - query
      title: TopicSearchRequest
      type: object
...
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          title: Location
          type: array
        msg:
          title: Message
          type: string
        type:
          title: Error Type
          type: string
      required:
      - loc
      - msg
      - type
      title: ValidationError
      type: object
  securitySchemes:
    ApiKeyAuth:
      in: header
      name: X-API-Key
      type: apiKey
    BearerAuth:
      bearerFormat: JWT
      scheme: bearer
      type: http
info:
  contact:
    email: mike@mikewaters.net
    name: LifeOS Development Team
    url: https://github.com/LifeOS/Substrate
  description: REST API for topic taxonomy and knowledge management
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  title: Ontology API
  version: 0.1.0
  x-api-id: ontology-api
  x-audience: internal
openapi: 3.1.0
paths:
...
  /topics:
    get:
      description: "List topics with optional filters and pagination.\n\nArgs:\n    service: Topic service dependency\n  \
        \  taxonomy_id: Filter by taxonomy\n    status_filter: Filter by status\n    created_after: Filter by creation date\
        \ (after)\n    created_before: Filter by creation date (before)\n    limit: Maximum number of results (1-100)\n  \
        \  offset: Number of results to skip\n\nReturns:\n    Paginated list of topics"
      operationId: list_topics_topics_get
      parameters:
      - description: Filter by taxonomy
        in: query
        name: taxonomy_id
        required: false
        schema:
          anyOf:
          - format: uuid
            type: string
          - type: 'null'
          description: Filter by taxonomy
          title: Taxonomy Id
      - description: Filter by status
        in: query
        name: status
        required: false
        schema:
          anyOf:
          - enum:
            - draft
            - active
            - deprecated
            - merged
            type: string
          - type: 'null'
          description: Filter by status
          title: Status
      - description: Filter by creation date (after)
        in: query
        name: created_after
        required: false
        schema:
          anyOf:
          - format: date-time
            type: string
          - type: 'null'
          description: Filter by creation date (after)
          title: Created After
      - description: Filter by creation date (before)
        in: query
        name: created_before
        required: false
        schema:
          anyOf:
          - format: date-time
            type: string
          - type: 'null'
          description: Filter by creation date (before)
          title: Created Before
      - description: Filter by materialized path prefix
        in: query
        name: path_prefix
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          description: Filter by materialized path prefix
          title: Path Prefix
      - in: query
        name: limit
        required: false
        schema:
          default: 50
          maximum: 100
          minimum: 1
          title: Limit
          type: integer
      - in: query
        name: offset
        required: false
        schema:
          default: 0
          minimum: 0
          title: Offset
          type: integer
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TopicListResponse'
              example:
                items:
                - &id002
                  id: 00000000-0000-0000-df70-3781f1070368
                  taxonomy_id: 00000000-0000-0000-d315-cb6cbdc35142
                  taxonomy_identifier: tx:tech
                  identifier: tech:software
                  title: Software
                  slug: software
                  description: Software development and engineering
                  status: active
                  path: /tech:software
                  aliases:
                  - software engineering
                  external_refs: {}
                  created_at: '2025-11-04T02:55:11.302186Z'
                  updated_at: '2025-11-04T02:55:11.302186Z'
                - id: 00000000-0000-0000-ebec-568259ed5d1b
                  taxonomy_id: 00000000-0000-0000-d315-cb6cbdc35142
                  taxonomy_identifier: tx:tech
                  identifier: tech:hardware
                  title: Hardware
                  slug: hardware
                  description: Physical computing components
                  status: active
                  path: /tech:hardware
                  aliases:
                  - electronics
                  external_refs: {}
                  created_at: '2025-11-04T02:55:11.302186Z'
                  updated_at: '2025-11-04T02:55:11.302186Z'
                total: 2
                limit: 50
                offset: 0
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: List topics with filters
      tags:
      - topics
    post:
      description: "Create a new topic.\n\nArgs:\n    topic: Topic creation data\n    service: Topic service dependency\n\n\
        Returns:\n    Created topic\n\nRaises:\n    ValidationError: If validation fails or slug is already taken"
      operationId: add_topic_topics_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TopicCreate'
        required: true
      responses:
        '201':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TopicResponse'
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Create a new topic
      tags:
      - topics
...
  /topics/multi-parent:
    get:
      description: "Find topics with multiple parent relationships.\n\nArgs:\n    service: Topic service dependency\n    min_parents:\
        \ Minimum number of parents\n    taxonomy_id: Optional filter by taxonomy\n\nReturns:\n    List of topics with multiple\
        \ parents"
      operationId: find_multi_parent_topics_topics_multi_parent_get
      parameters:
      - description: Minimum number of parents
        in: query
        name: min_parents
        required: false
        schema:
          default: 2
          description: Minimum number of parents
          minimum: 2
          title: Min Parents
          type: integer
      - description: Filter by taxonomy
        in: query
        name: taxonomy_id
        required: false
        schema:
          anyOf:
          - format: uuid
            type: string
          - type: 'null'
          description: Filter by taxonomy
          title: Taxonomy Id
      responses:
        '200':
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/TopicResponse'
                title: Response Find Multi Parent Topics Topics Multi Parent Get
                type: array
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Find topics with multiple parents
      tags:
      - topics
  /topics/orphans:
    get:
      description: "Find topics with no parent relationships.\n\nArgs:\n    service: Topic service dependency\n    taxonomy_id:\
        \ Optional filter by taxonomy\n\nReturns:\n    List of orphan topics"
      operationId: find_orphan_topics_topics_orphans_get
      parameters:
      - description: Filter by taxonomy
        in: query
        name: taxonomy_id
        required: false
        schema:
          anyOf:
          - format: uuid
            type: string
          - type: 'null'
          description: Filter by taxonomy
          title: Taxonomy Id
      responses:
        '200':
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/TopicResponse'
                title: Response Find Orphan Topics Topics Orphans Get
                type: array
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Find orphan topics
      tags:
      - topics
  /topics/search:
    post:
      description: "Search for topics by title or alias.\n\nSupports case-insensitive substring matching with optional filters\n\
        for taxonomy and status.\n\nArgs:\n    request: Search request with query and filters\n    service: Topic service\
        \ dependency\n\nReturns:\n    Paginated list of matching topics"
      operationId: search_topics_topics_search_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TopicSearchRequest'
        required: true
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TopicListResponse'
              example:
                items:
                - *id002
                total: 1
                limit: 50
                offset: 0
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Search for topics
      tags:
      - topics
...
servers:
- description: Development server
  url: http://localhost:8000
- description: Production server
  url: https://api.lifeos.dev